+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 * #18 bootique-kafka-streams: a new module to support Kafka streams
 * #19 Rename the project to "bootique-kafka"
 * #21 "kafkaclient.producer.lingerMs" - rename to "linger" and use Duration type
+* #22 Support for "auto.offset.reset" property for consumers
 
 ## 0.25